Refactor into multiple rust binaries
