+++ ping/rust/Cargo.toml
@@ -4,7 +4,9 @@ name = "testplan"
 version = "0.1.0"
 
 [dependencies]
-async-std = {version = "1.10", features = ["attributes", "tokio1"]}
+async-trait = "0.1.58"
+anyhow = "1"
+async-std = { version = "1.10", features = ["attributes", "tokio1"] }
 env_logger = "0.9.0"
 if-addrs = "0.7.0"
 log = "0.4"
@@ -18,11 +20,12 @@ tokio-stream = { version = "0.1", default-features = false, features = [] }
 tokio-util = { version = "0.7", default-features = false, features = ["compat"] }
 futures = "0.3.1"
 
-libp2pv0440 = {package = "libp2p", default_features = false, features = [ "websocket", "mplex", "yamux", "tcp-async-io", "ping", "noise", "dns-async-std" ], version = "0.44.0", optional = true}
-libp2pv0450 = {package = "libp2p", default_features = false, features = [ "websocket", "mplex", "yamux", "tcp-async-io", "ping", "noise", "dns-async-std" ], version = "0.45.0", optional = true}
-libp2pv0460 = {package = "libp2p", default_features = false, features = [ "websocket", "mplex", "yamux", "tcp-async-io", "ping", "noise", "dns-async-std" ], version = "0.46.0", optional = true}
-libp2pv0470 = {package = "libp2p", default_features = false, features = [ "websocket", "mplex", "yamux", "tcp-async-io", "ping", "noise", "dns-async-std" ], version = "0.47.0", optional = true}
-libp2pv0480 = {package = "libp2p", default_features = false, features = [ "websocket", "mplex", "yamux", "tcp-async-io", "ping", "noise", "dns-async-std", "rsa" ], version = "0.48.0", optional = true}
-libp2pv0490 = {package = "libp2p", default_features = false, features = [ "websocket", "mplex", "yamux", "tcp-async-io", "ping", "noise", "dns-async-std", "rsa" ], version = "0.49.0", optional = true}
+libp2pv0440 = { package = "libp2p", version = "0.44.0", default_features = false, features = ["websocket", "mplex", "yamux", "tcp-async-io", "ping", "noise", "dns-async-std"] }
+libp2pv0450 = { package = "libp2p", version = "0.45.0", default_features = false, features = ["websocket", "mplex", "yamux", "tcp-async-io", "ping", "noise", "dns-async-std"] }
+libp2pv0460 = { package = "libp2p", version = "0.46.0", default_features = false, features = ["websocket", "mplex", "yamux", "tcp-async-io", "ping", "noise", "dns-async-std"] }
+libp2pv0470 = { package = "libp2p", version = "0.47.0", default_features = false, features = ["websocket", "mplex", "yamux", "tcp-async-io", "ping", "noise", "dns-async-std"] }
+libp2pv0480 = { package = "libp2p", version = "0.48.0", default_features = false, features = ["websocket", "mplex", "yamux", "tcp-async-io", "ping", "noise", "dns-async-std", "rsa"] }
+libp2pv0490 = { package = "libp2p", version = "0.49.0", features = ["websocket", "mplex", "yamux", "tcp", "async-std", "ping", "noise", "dns", "rsa"] }
 
-libp2pmaster = {package = "libp2p", git = "https://github.com/libp2p/rust-libp2p", branch = "master", default_features = false, features = [ "websocket", "mplex", "yamux", "tcp", "ping", "noise", "dns", "async-std", "rsa" ], optional = true}
+# Next release
+libp2pv0500 = { package = "libp2p", git = "https://github.com/libp2p/rust-libp2p", features = ["websocket", "mplex", "yamux", "tcp", "ping", "noise", "dns", "async-std", "rsa"] }
